# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SubastApp API",
    "description": "API documentation for SubastApp",
    "license": {
      "name": "FabrizioB Website",
      "url": "https://fabriziob.com/"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4002",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/personas/{id}": {
      "get": {
        "tags": [
          "persona-controller"
        ],
        "operationId": "getById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "persona-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "persona-controller"
        ],
        "operationId": "desactivar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/empleados": {
      "get": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "getAll_4",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageEmpleadoResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/empleados/{id}": {
      "get": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "getById_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "desactivar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clientes/{id}/medios-pago/{mpId}/desactivar": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Desactivar medio de pago",
        "description": "Desactiva un medio de pago, impidiendo su uso en futuras operaciones",
        "operationId": "desactivar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "mpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Medio de pago desactivado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente o medio de pago no encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PersonaUpdateRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "foto": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "PersonaResponse": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          }
        }
      },
...
      "EmpleadoRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "cargo": {
            "type": "string"
          },
          "sector": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EmpleadoResponse": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "cargo": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "sectorId": {
            "type": "integer",
            "format": "int32"
          },
          "sectorNombre": {
            "type": "string"
          }
        }
      },
...
      "Moneda": {
        "type": "string",
        "enum": [
          "ARS",
          "USD"
        ]
      },
      "MedioPagoChequeDetail": {
        "type": "object",
        "properties": {
          "nroCheque": {
            "type": "string"
          },
          "banco": {
            "type": "string"
          },
          "montoCertificado": {
            "type": "number"
          },
          "fechaVencimiento": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "MedioPagoCuentaDetail": {
        "type": "object",
        "properties": {
          "titular": {
            "type": "string"
          },
          "banco": {
            "type": "string"
          },
          "cbu": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "esExterior": {
            "type": "boolean"
          },
          "iban": {
            "type": "string"
          },
          "pais": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MedioPagoResponse": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "integer",
            "format": "int32"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoMedioPago"
          },
          "moneda": {
            "$ref": "#/components/schemas/Moneda"
          },
          "verificado": {
            "type": "boolean"
          },
          "activo": {
            "type": "boolean"
          },
          "cheque": {
            "$ref": "#/components/schemas/MedioPagoChequeDetail"
          },
          "cuenta": {
            "$ref": "#/components/schemas/MedioPagoCuentaDetail"
          },
          "tarjeta": {
            "$ref": "#/components/schemas/MedioPagoTarjetaDetail"
          }
        }
      },
      "MedioPagoTarjetaDetail": {
        "type": "object",
        "properties": {
          "titular": {
            "type": "string"
          },
          "ultimos4": {
            "type": "string"
          },
          "marca": {
            "type": "string"
          },
          "vencimiento": {
            "type": "string"
          },
          "esInternacional": {
            "type": "boolean"
          }
        }
      },
      "TipoMedioPago": {
        "type": "string",
        "enum": [
          "cheque",
          "cuenta_bancaria",
          "tarjeta_credito"
        ]
      },
...
      "EmpleadoUpdateRequest": {
        "type": "object",
        "properties": {
          "cargo": {
            "type": "string"
          },
          "sector": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageEmpleadoResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmpleadoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}